Reused orphaned variable StringKick to model effect discussed by http://arxiv.org/abs/1005.2416 Just specify the velocity in km/s that the baryons should move relative to the DM at the initial redshift.
Also added the original Truelove gaussian sphere
collapse to th CollapseMHD3DInitialize test suite.

--HG--
branch : euler_improve

+++ src/enzo/global_data.h
@@ -657,7 +657,7 @@ EXTERN int UseConstantAcceleration;
 EXTERN float ConstantAcceleration[3];
 EXTERN float Mu;
 EXTERN int ExternalGravity;
-EXTERN int StringKick;
+EXTERN float StringKick;
 EXTERN int UseFloor;
 EXTERN int UseViscosity;
 EXTERN float ViscosityCoefficient;
